pyproject.toml
[project]
name = "forecast-sales-lm"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "jinja2>=3.1.4",
    "numpy>=2.0.2",
    "pandas>=2.2.3",
    "pydantic>=2.10.2",
    "pyyaml>=6.0.2",
]
